+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This is the changelog for Tarmac.
 
 ## [Unreleased]
 
+### Added
+
+- Allow multiple commands to be specified in a `run` workflow step
+
 ## [0.1.6]
 
 ### Added